requirements.txt
asgiref==3.8.1
boto3==1.34.143
botocore==1.34.143
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
moto==5.0.11
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
responses==0.25.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
Werkzeug==3.0.3
xmltodict==0.13.0
